Move SB diagnostics to a module
